fix ruff/mypy/tsc lint findings in engine and fresh-route
marseille-engine: sort imports, drop unused imports, Optional[X] -> X | None,
zip() -> itertools.pairwise(), drop redundant int() around round(), narrow
uhi_to_label's return type to the actual Literal, fix a real type bug in
timeline.py (was reading a possibly-None dict value instead of the
already-correctly-typed anomaly float). Added mypy overrides for geo libs
with no stubs (rasterio/geopandas/rasterstats) and installed stubs for
pandas/PyYAML/networkx. The two blind `except Exception` in route.py are a
deliberate catch-all -> annotated with noqa instead of narrowed.

fresh-route: Route3DPin.tsx imported a nonexistent `Text` from
@react-three/fiber and pulled in useRef via a bottom-of-file require()
that typed as `any`, breaking useRef<THREE.Mesh>(); RouteMap.tsx used
StyleSheet.absoluteFillObject, renamed to absoluteFill in RN 0.85.

ruff check, mypy, and tsc --noEmit all pass clean now.

+++ marseille-engine/api/routers/raster.py
@@ -1,8 +1,9 @@
 import numpy as np
-from fastapi import APIRouter, Request, Query
-from fastapi.responses import Response, JSONResponse
+from fastapi import APIRouter, Query, Request
+from fastapi.responses import JSONResponse, Response
+
+from api.routers.timeline import BASELINE_END, BASELINE_START, ZONES, _zone_annual
 from core.raster_renderer import raster_to_png
-from api.routers.timeline import _zone_annual, ZONES, BASELINE_START, BASELINE_END
 
 router = APIRouter()
 
@@ -13,7 +14,7 @@ def get_raster_png(request: Request, scale: float = Query(0.25, ge=0.05, le=1.0)
     loader = request.app.state.loader
     data = loader.raster_data.copy()
     mask = np.isnan(data)
-    step = max(1, int(round(1.0 / scale)))
+    step = max(1, round(1.0 / scale))
     png_bytes = raster_to_png(data[::step, ::step], mask[::step, ::step])
     return Response(content=png_bytes, media_type="image/png", headers={
         "Cache-Control": "public, max-age=3600",
@@ -30,7 +31,7 @@ def get_raster_raw(request: Request, scale: float = Query(0.3, ge=0.05, le=1.0))
     """
     loader = request.app.state.loader
     data = loader.raster_data  # already NaN for nodata
-    step = max(1, int(round(1.0 / scale)))
+    step = max(1, round(1.0 / scale))
     downsampled = data[::step, ::step].astype(np.float32)
 
     rows, cols = downsampled.shape
